use only opentelemetry in dskit
