+++ spanlogger/spanlogger_test.go
@@ -6,17 +6,24 @@ import (
 	"testing"
 
 	"github.com/go-kit/log"
-	"github.com/opentracing/opentracing-go"
-	"github.com/opentracing/opentracing-go/mocktracer"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/user"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 )
 
-func TestSpanLogger_Log(t *testing.T) {
+func TestOtelSpanLogger_Log(t *testing.T) {
 	logger := log.NewNopLogger()
 	resolver := fakeResolver{}
+	tp := sdktrace.NewTracerProvider(
+		sdktrace.WithBatcher(tracetest.NewInMemoryExporter()),
+	)
+	otel.SetTracerProvider(tp)
 	span, ctx := New(context.Background(), logger, "test", resolver, "bar")
+	defer span.End()
 	_ = span.Log("foo")
 	newSpan := FromContext(ctx, logger, resolver)
 	require.Equal(t, span.Span, newSpan.Span)
@@ -27,13 +34,20 @@ func TestSpanLogger_Log(t *testing.T) {
 	require.NoError(t, noSpan.Error(nil))
 }
 
-func TestSpanLogger_CustomLogger(t *testing.T) {
+func TestOtelSpanLogger_CustomLogger(t *testing.T) {
 	var logged [][]interface{}
 	var logger funcLogger = func(keyvals ...interface{}) error {
 		logged = append(logged, keyvals)
 		return nil
 	}
 	resolver := fakeResolver{}
+	exp := tracetest.NewNoopExporter()
+	tp := sdktrace.NewTracerProvider(
+		sdktrace.WithBatcher(exp),
+		// Set the sampler to never sample so that traceID is not logged.
+		sdktrace.WithSampler(sdktrace.NeverSample()),
+	)
+	otel.SetTracerProvider(tp)
 	span, ctx := New(context.Background(), logger, "test", resolver)
 	_ = span.Log("msg", "original spanlogger")
 
@@ -48,28 +62,54 @@ func TestSpanLogger_CustomLogger(t *testing.T) {
 		{"msg", "restored spanlogger"},
 		{"msg", "fallback spanlogger"},
 	}
+
 	require.Equal(t, expect, logged)
 }
 
-func TestSpanCreatedWithTenantTag(t *testing.T) {
-	mockSpan := createSpan(user.InjectOrgID(context.Background(), "team-a"))
+func TestOtelSpanCreatedWithTenantTag(t *testing.T) {
+	exp, sp := createOtelSpan(user.InjectOrgID(context.Background(), "team-a"))
+	defer sp.End()
 
-	require.Equal(t, []string{"team-a"}, mockSpan.Tag(TenantIDsTagName))
+	require.Equal(t, 1, len(exp.GetSpans().Snapshots()))
+	require.Equal(t,
+		[]attribute.KeyValue{attribute.StringSlice(TenantIDsTagName, []string{"team-a"})},
+		exp.GetSpans().Snapshots()[0].Attributes())
 }
 
-func TestSpanCreatedWithoutTenantTag(t *testing.T) {
-	mockSpan := createSpan(context.Background())
+func TestOtelSpanCreatedWithoutTenantTag(t *testing.T) {
+	exp, sp := createOtelSpan(context.Background())
+	defer sp.End()
+	require.Equal(t, 1, len(exp.GetSpans().Snapshots()))
 
-	_, exists := mockSpan.Tags()[TenantIDsTagName]
-	require.False(t, exists)
+	exist := false
+	for _, kv := range exp.GetSpans().Snapshots()[0].Attributes() {
+		if kv.Key == TenantIDsTagName {
+			exist = true
+		}
+	}
+	require.False(t, exist)
 }
 
-func createSpan(ctx context.Context) *mocktracer.MockSpan {
-	mockTracer := mocktracer.New()
-	opentracing.SetGlobalTracer(mockTracer)
+// Using a no-op logger and no tracing provider, measure the overhead of a small log call.
+func BenchmarkOtelSpanLogger(b *testing.B) {
+	_, sl := createOtelSpan(context.Background())
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_ = sl.Log("msg", "foo", "more", "data")
+	}
+}
 
-	logger, _ := New(ctx, log.NewNopLogger(), "name", fakeResolver{})
-	return logger.Span.(*mocktracer.MockSpan)
+func createOtelSpan(ctx context.Context) (*tracetest.InMemoryExporter, *SpanLogger) {
+	exp := tracetest.NewInMemoryExporter()
+	tp := sdktrace.NewTracerProvider(
+		sdktrace.WithBatcher(exp),
+	)
+	otel.SetTracerProvider(tp)
+
+	sl, _ := New(ctx, log.NewNopLogger(), "get", fakeResolver{})
+	// Force flush to ensure spans are reported before the test ends.
+	tp.ForceFlush(ctx)
+	return exp, sl
 }
 
 type funcLogger func(keyvals ...interface{}) error
@@ -105,15 +145,4 @@ func (r fakeResolver) TenantIDs(ctx context.Context) ([]string, error) {
 	}
 
 	return []string{id}, nil
-}
-
-// Using a no-op logger and no tracing provider, measure the overhead of a small log call.
-func BenchmarkSpanLogger(b *testing.B) {
-	logger := log.NewNopLogger()
-	resolver := fakeResolver{}
-	sl, _ := New(context.Background(), logger, "test", resolver, "bar")
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		_ = sl.Log("msg", "foo", "more", "data")
-	}
 }
